+++ modules/admin-ui-index/src/main/resources/OSGI-INF/message-receiver-acl.xml
@@ -30,7 +30,7 @@
 
   <reference name="search-index"
              cardinality="1..1"
-             interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+             interface="org.opencastproject.adminui.index.AdminUISearchIndex"
              policy="static"
              bind="setSearchIndex"/>
