+++ modules/admin-ui-index/src/main/resources/OSGI-INF/message-receiver-scheduler.xml
@@ -16,7 +16,7 @@
   <reference name="message-receiver-lock-service"
              interface="org.opencastproject.index.service.message.MessageReceiverLockService"
              cardinality="1..1" policy="static" bind="setMessageReceiverLockService"/>
-  <reference name="search-index" interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+  <reference name="search-index" interface="org.opencastproject.adminui.index.AdminUISearchIndex"
              cardinality="1..1" policy="static" bind="setSearchIndex"/>
   <reference name="security-service" interface="org.opencastproject.security.api.SecurityService"
              cardinality="1..1" policy="static" bind="setSecurityService"/>
