+++ modules/admin-ui-index/src/main/resources/OSGI-INF/message-receiver-theme.xml
@@ -1,33 +1,43 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0"
-               name="org.opencastproject.index.service.message.WorkflowMessageReceiverImpl-Adminui"
-               immediate="true">
+               name="org.opencastproject.index.service.message.ThemeMessageReceiverImpl-Adminui"
+               immediate="true" activate="activate" deactivate="deactivate">
   <implementation
-    class="org.opencastproject.index.service.message.WorkflowMessageReceiverImpl"/>
-  <property name="service.description" value="Workflow Message Receiver"/>
-  <property name="destinationId" value="WORKFLOW.Adminui"/>
+    class="org.opencastproject.index.service.message.ThemeMessageReceiverImpl"/>
+  <property name="service.description" value="Theme Message Receiver"/>
+  <property name="destinationId" value="THEME.Adminui"/>
   <service>
-    <provide interface="org.opencastproject.index.service.message.WorkflowMessageReceiverImpl"/>
+    <provide interface="org.opencastproject.index.service.message.ThemeMessageReceiverImpl"/>
   </service>
 
   <reference name="message-broker-sender"
+             cardinality="1..1"
              interface="org.opencastproject.message.broker.api.MessageSender"
+             policy="static"
              bind="setMessageSender"/>
 
   <reference name="message-broker-receiver"
+             cardinality="1..1"
              interface="org.opencastproject.message.broker.api.MessageReceiver"
+             policy="static"
              bind="setMessageReceiver"/>
 
   <reference name="message-receiver-lock-service"
+             cardinality="1..1"
              interface="org.opencastproject.index.service.message.MessageReceiverLockService"
+             policy="static"
              bind="setMessageReceiverLockService"/>
 
   <reference name="search-index"
-             interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+             cardinality="1..1"
+             interface="org.opencastproject.adminui.index.AdminUISearchIndex"
+             policy="static"
              bind="setSearchIndex"/>
 
   <reference name="security-service"
              interface="org.opencastproject.security.api.SecurityService"
+             cardinality="1..1"
+             policy="static"
              bind="setSecurityService"/>
 
 </scr:component>
