+++ modules/admin-ui-index/src/main/resources/OSGI-INF/message-receiver-workflow.xml
@@ -1,43 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0"
-               name="org.opencastproject.index.service.message.ThemeMessageReceiverImpl-Adminui"
-               immediate="true" activate="activate" deactivate="deactivate">
+               name="org.opencastproject.index.service.message.WorkflowMessageReceiverImpl-Adminui"
+               immediate="true">
   <implementation
-    class="org.opencastproject.index.service.message.ThemeMessageReceiverImpl"/>
-  <property name="service.description" value="Theme Message Receiver"/>
-  <property name="destinationId" value="THEME.Adminui"/>
+    class="org.opencastproject.index.service.message.WorkflowMessageReceiverImpl"/>
+  <property name="service.description" value="Workflow Message Receiver"/>
+  <property name="destinationId" value="WORKFLOW.Adminui"/>
   <service>
-    <provide interface="org.opencastproject.index.service.message.ThemeMessageReceiverImpl"/>
+    <provide interface="org.opencastproject.index.service.message.WorkflowMessageReceiverImpl"/>
   </service>
 
   <reference name="message-broker-sender"
-             cardinality="1..1"
              interface="org.opencastproject.message.broker.api.MessageSender"
-             policy="static"
              bind="setMessageSender"/>
 
   <reference name="message-broker-receiver"
-             cardinality="1..1"
              interface="org.opencastproject.message.broker.api.MessageReceiver"
-             policy="static"
              bind="setMessageReceiver"/>
 
   <reference name="message-receiver-lock-service"
-             cardinality="1..1"
              interface="org.opencastproject.index.service.message.MessageReceiverLockService"
-             policy="static"
              bind="setMessageReceiverLockService"/>
 
   <reference name="search-index"
-             cardinality="1..1"
-             interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
-             policy="static"
+             interface="org.opencastproject.adminui.index.AdminUISearchIndex"
              bind="setSearchIndex"/>
 
   <reference name="security-service"
              interface="org.opencastproject.security.api.SecurityService"
-             cardinality="1..1"
-             policy="static"
              bind="setSecurityService"/>
 
 </scr:component>
