+++ modules/admin-ui-index/src/test/java/org/opencastproject/adminui/index/AdminUISearchIndexStub.java
@@ -20,7 +20,7 @@
  */
 
 
-package org.opencastproject.adminui.impl.index;
+package org.opencastproject.adminui.index;
 
 import org.opencastproject.matterhorn.search.SearchIndexException;
