+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/endpoint/IndexEndpoint.java
@@ -21,7 +21,7 @@
 
 package org.opencastproject.adminui.endpoint;
 
-import org.opencastproject.adminui.impl.index.AdminUISearchIndex;
+import org.opencastproject.adminui.index.AdminUISearchIndex;
 import org.opencastproject.security.api.SecurityService;
 import org.opencastproject.security.util.SecurityContext;
 import org.opencastproject.util.RestUtil.R;
