+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/util/BulkUpdateUtil.java
@@ -25,7 +25,7 @@ import static org.opencastproject.adminui.endpoint.AbstractEventEndpoint.SCHEDUL
 import static org.opencastproject.adminui.endpoint.AbstractEventEndpoint.SCHEDULING_END_KEY;
 import static org.opencastproject.adminui.endpoint.AbstractEventEndpoint.SCHEDULING_START_KEY;
 
-import org.opencastproject.adminui.impl.index.AdminUISearchIndex;
+import org.opencastproject.adminui.index.AdminUISearchIndex;
 import org.opencastproject.index.service.api.IndexService;
 import org.opencastproject.index.service.catalog.adapter.events.CommonEventCatalogUIAdapter;
 import org.opencastproject.index.service.impl.index.event.Event;
