+++ modules/admin-ui/src/main/resources/OSGI-INF/event_endpoint.xml
@@ -13,68 +13,42 @@
     <provide interface="org.osgi.service.cm.ManagedService"/>
   </service>
   <reference bind="setAdminUIConfiguration"
-             cardinality="1..1"
              interface="org.opencastproject.adminui.impl.AdminUIConfiguration"
-             name="adminUIConfiguration"
-             policy="static"/>
+             name="adminUIConfiguration"/>
   <reference bind="setWorkflowService"
-             cardinality="1..1"
              interface="org.opencastproject.workflow.api.WorkflowService"
-             name="workflowService"
-             policy="static"/>
+             name="workflowService"/>
   <reference name="jobService"
              interface="org.opencastproject.adminui.endpoint.JobEndpoint"
-             cardinality="1..1"
-             policy="static"
              bind="setJobService"/>
-  <reference name="seriesService"
+  <reference name="seriesEndpoint"
              interface="org.opencastproject.adminui.endpoint.SeriesEndpoint"
-             cardinality="1..1"
-             policy="static"
              bind="setSeriesService"/>
   <reference name="AclServiceFactory"
              interface="org.opencastproject.authorization.xacml.manager.api.AclServiceFactory"
-             cardinality="1..1"
-             policy="static"
              bind="setAclServiceFactory"/>
   <reference name="EventCommentService"
              interface="org.opencastproject.event.comment.EventCommentService"
-             cardinality="1..1"
-             policy="static"
              bind="setEventCommentService"/>
   <reference name="SecurityService"
              interface="org.opencastproject.security.api.SecurityService"
-             cardinality="1..1"
-             policy="static"
              bind="setSecurityService"/>
   <reference name="AuthorizationService"
              interface="org.opencastproject.security.api.AuthorizationService"
-             cardinality="1..1"
-             policy="static"
              bind="setAuthorizationService"/>
   <reference name="SchedulerService"
              interface="org.opencastproject.scheduler.api.SchedulerService"
-             cardinality="1..1"
-             policy="static"
              bind="setSchedulerService"/>
   <reference name="CaptureAgentStateService"
              interface="org.opencastproject.capture.admin.api.CaptureAgentStateService"
-             cardinality="1..1"
-             policy="static"
              bind="setCaptureAgentStateService"/>
   <reference name="IndexService"
              interface="org.opencastproject.index.service.api.IndexService"
-             cardinality="1..1"
-             policy="static"
              bind="setIndexService"/>
   <reference name="AdminUISearchIndex"
-             interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
-             cardinality="1..1"
-             policy="static"
+             interface="org.opencastproject.adminui.index.AdminUISearchIndex"
              bind="setIndex"/>
   <reference name="UrlSigningService"
              interface="org.opencastproject.security.urlsigning.service.UrlSigningService"
-             cardinality="1..1"
-             policy="static"
              bind="setUrlSigningService"/>
 </scr:component>
