+++ modules/admin-ui/src/main/resources/OSGI-INF/groups_endpoint.xml
@@ -17,6 +17,6 @@
              cardinality="1..1" policy="static" bind="setUserDirectoryService"/>
   <reference name="IndexService" interface="org.opencastproject.index.service.api.IndexService"
              cardinality="1..1" policy="static" bind="setIndexService"/>
-  <reference name="AdminUISearchIndex" interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+  <reference name="AdminUISearchIndex" interface="org.opencastproject.adminui.index.AdminUISearchIndex"
              cardinality="1..1" policy="static" bind="setSearchIndex"/>
 </scr:component>
