--- modules/admin-ui/src/main/resources/OSGI-INF/search-index.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0"
-               immediate="true"
-               activate="activate"
-               deactivate="deactivate"
-               name="org.opencastproject.adminui.impl.index.AdminUISearchIndex">
-
-  <implementation class="org.opencastproject.adminui.impl.index.AdminUISearchIndex"/>
-  <property name="service.description" value="org.opencastproject.adminui.impl.index.AdminUISearchIndex"/>
-  <property name="service.pid" value="ch.entwine.matterhorn.search"/>
-  <service>
-    <provide interface="org.opencastproject.matterhorn.search.SearchIndex"/>
-    <provide interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"/>
-  </service>
-
-  <reference name="MessageSender" interface="org.opencastproject.message.broker.api.MessageSender"
-             cardinality="1..1" policy="static" bind="setMessageSender"/>
-  <reference name="MessageReceiver" interface="org.opencastproject.message.broker.api.MessageReceiver"
-             cardinality="1..1" policy="static" bind="setMessageReceiver"/>
-</scr:component>
