+++ modules/admin-ui/src/main/resources/OSGI-INF/series_endpoint.xml
@@ -13,7 +13,7 @@
 
   <reference name="AclServiceFactory" interface="org.opencastproject.authorization.xacml.manager.api.AclServiceFactory"
              cardinality="1..1" policy="static" bind="setAclServiceFactory"/>
-  <reference name="AdminUISearchIndex" interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+  <reference name="AdminUISearchIndex" interface="org.opencastproject.adminui.index.AdminUISearchIndex"
              cardinality="1..1" policy="static" bind="setIndex"/>
   <reference name="IndexService" interface="org.opencastproject.index.service.api.IndexService"
              cardinality="1..1" policy="static" bind="setIndexService"/>
