+++ modules/admin-ui/src/main/resources/OSGI-INF/themes_endpoint.xml
@@ -14,7 +14,7 @@
              cardinality="1..1" policy="static" bind="setThemesServiceDatabase"/>
   <reference name="security-service" interface="org.opencastproject.security.api.SecurityService"
              cardinality="1..1" policy="static" bind="setSecurityService"/>
-  <reference name="AdminUISearchIndex" interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+  <reference name="AdminUISearchIndex" interface="org.opencastproject.adminui.index.AdminUISearchIndex"
              cardinality="1..1" policy="static" bind="setIndex"/>
   <reference name="seriesService" interface="org.opencastproject.series.api.SeriesService"
              cardinality="1..1" policy="static" bind="setSeriesService"/>
