+++ modules/index-service/src/main/resources/OSGI-INF/list-providers/contributors.xml
@@ -18,7 +18,7 @@
              bind="setUserDirectoryService"/>
 
   <reference name="AdminUISearchIndex"
-             interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+             interface="AdminUISearchIndex"
              cardinality="1..1"
              policy="static"
              bind="setIndex"/>
