+++ modules/index-service/src/main/resources/OSGI-INF/list-providers/themes.xml
@@ -12,7 +12,7 @@
   </service>
 
   <reference name="AdminUISearchIndex"
-             interface="org.opencastproject.adminui.impl.index.AdminUISearchIndex"
+             interface="AdminUISearchIndex"
              cardinality="1..1"
              policy="static"
              bind="setIndex"/>
