+++ pom.xml
@@ -46,6 +46,7 @@
   </properties>
   <modules>
     <module>modules/admin-ui</module>
+    <module>modules/admin-ui-index</module>
     <module>modules/animate-api</module>
     <module>modules/animate-impl</module>
     <module>modules/animate-remote</module>
